+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Add a Category ScheduleOnSheet and its nodes - ScheduleOnSheet.Sheet, ScheduleOnSheet.Schedule, ScheduleOnSheet.BySheetViewLocation, ScheduleOnSheet.Location and ScheduleOnSheet.SetLocation
+
 ## 0.2.17
 * Check if InternalElement is valid Revit Element to avoid calling funcitons or get properties of an invalid object and cause Dynamo crash.
 * Add some new nodes to Viewport - Viewport.Sheet, Viewport.View, Viewport.BoxCenter, Viewport.SetBoxCenter, Viewport.BoxOutline and Viewport.LabelOutline.